requirements.txt
langgraph
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.3.2
langchain
llama-index
xgboost
fastapi
uvicorn
python-dotenv
langchain-groq